Added Model Layer
Created a model layer responsible for retrieving data.
